refactor: #105 - 링크 생성 훅과 메타 정보를 가져오는 훅을 분리
